*: move HashMap and HashSet to util::collections
