+++ src/server/config.rs
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use util::HashMap;
+use util::collections::HashMap;
 
 use super::Result;
