+++ src/server/resolve.rs
@@ -19,7 +19,8 @@ use std::time::Instant;
 
 use kvproto::metapb;
 
-use util::{self, HashMap};
+use util;
+use util::collections::HashMap;
 use util::worker::{Runnable, Worker};
 use pd::PdClient;
 
@@ -157,7 +158,8 @@ mod tests {
     use kvproto::pdpb;
     use kvproto::metapb;
     use pd::{PdClient, Result};
-    use util::{self, HashMap};
+    use util;
+    use util::collections::HashMap;
 
     const STORE_ADDRESS_REFRESH_SECONDS: u64 = 60;
