+++ tests/raft/test_raft.rs
@@ -37,7 +37,7 @@ use rand;
 
 use tikv::raft::*;
 use tikv::raft::storage::MemStorage;
-use tikv::util::{HashMap as RaftHashMap, BuildHasherDefault};
+use tikv::util::collections::{HashMap as RaftHashMap, BuildHasherDefault};
 
 pub fn ltoa(raft_log: &RaftLog<MemStorage>) -> String {
     let mut s = format!("committed: {}\n", raft_log.committed);
